rename all the things
This should cover ever instance of "nullshield" in the project.
In addition it renames the project for npm under the @iwburns scope.
