+++ rollup.config.ts
@@ -7,7 +7,7 @@ import json from 'rollup-plugin-json'
 
 const pkg = require('./package.json')
 
-const libraryName = 'nullshield'
+const libraryName = 'tupperware'
 
 export default {
   input: `src/${libraryName}.ts`,
